Redesign of Home page beginning
Added a plain background with a simple patter, debating how to add more colour as home looks bland at the moment. Furthermore, I am deciding whether or not I am keeping this image for the home page -> No decisions have been made yet. Also, added an animation to the image on the homepage.